requirements.txt
opencv-python
numpy
pyautogui
pillow
opencv-python-headless
pytesseract
chroma_brain
selenium
webdriver-manager